+++ app_Main.py
@@ -629,7 +629,7 @@ class App(QtCore.QObject):
                                   'export_gcode', 'export_gerber', 'export_svg', 'ext', 'exteriors', 'follow',
                                   'geo_union', 'geocutout', 'get_bounds', 'get_names', 'get_path', 'get_sys', 'help',
                                   'interiors', 'isolate', 'join_excellon',
-                                  'join_geometry', 'list_sys', 'milld', 'mills', 'milldrills', 'millslots',
+                                  'join_geometry', 'list_sys', 'list_pp', 'milld', 'mills', 'milldrills', 'millslots',
                                   'mirror', 'ncc',
                                   'ncr', 'new', 'new_geometry', 'new_gerber', 'new_excellon', 'non_copper_regions',
                                   'offset',